requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
rawpy>=0.19.0
requests>=2.31.0
